% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senseBox_archive.R
\name{get_senseBox_archive}
\alias{get_senseBox_archive}
\title{Get archived data from a senseBox}
\usage{
get_senseBox_archive(senseBoxId, date = NULL)
}
\arguments{
\item{senseBoxId}{\link{character} (\strong{required}): senseBoxId}

\item{date}{\link{character} (\strong{optional}): Download data from date in format YYYY-mm-dd. If no date was given,
the functions tries to download the archive from yesterday.}
}
\value{
The function returns an \link[httr:response]{httr::response} object.
}
\description{
Get archived data from a senseBox
}
\section{Function version}{
 0.0.1
}

\examples{

##=====================================
## Example: download senseBox archive
##=====================================

get_senseBox_archive(
  senseBoxId = "5957b67494f0520011304cc1",
  date = "2018-01-01")

get_senseBox_archive(
  senseBoxId = "5957b67494f0520011304cc1")

}
\author{
Johannes Friedrich
}
