% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senseBox_data.R
\name{get_senseBox_data}
\alias{get_senseBox_data}
\title{Get data from one specific senseBox sensor.}
\usage{
get_senseBox_data(senseBoxId, sensorId = rep("all", times =
  length(unlist(senseBoxId))), fromDate = NULL, toDate = NULL,
  CSV = FALSE, POSIXct = TRUE)
}
\arguments{
\item{senseBoxId}{\link{character}or \link{vector} of \link{character} (\strong{required}): senseBoxId or a \link{vector} of \link{character} of senseBoxIds}

\item{sensorId}{\link{character} or \link{list} of \link{character} (\strong{required}): sensorId or a \link{list} of senseBoxIds}

\item{fromDate}{\link{character} (\strong{optional}): Just show data with from date in format YYYY-mm-dd HH:MM:SS}

\item{toDate}{\link{character} (\strong{optional}): Just show data with up to date in format YYYY-mm-dd HH:MM:SS}

\item{CSV}{\link{logical} (\strong{optional}): Download data as .csv? NOT POSSIBLE UNTIL NOW!}

\item{POSIXct}{\link{logical} (\strong{optional}): Should the timestamp be translated into POSIXct?}
}
\value{
\link{list}
}
\description{
If the arguments \code{fromDate} and \code{toDate} are NOT given, the measurements from the last 48 h from
now are downloaded. The maximum numbers of downloaded measurements are 10,000.
The maximum time frame is one month.
It is possible to download the data directly as a csv file, when argument \code{CSV} is \code{TRUE}.
}
\section{Function version}{
 0.0.1
}

\examples{

##=====================================
## Example: get data from sensorId
##=====================================

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4")

\dontrun{
get_senseBox_data(
  senseBoxId = c("5957b67494f0520011304cc1","5957b67494f0520011304cc1") ,
  sensorId = "all",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")

get_senseBox_data(
  senseBoxId = "5957b67494f0520011304cc1",
  sensorId = "5957b67494f0520011304cc4",
  fromDate = "2017-11-25-12:00:00",
  toDate = "2017-11-26-12:00:00")
}

}
\author{
Johannes Friedrich
}
