% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senseBox_Ids.R
\name{get_senseBox_Ids}
\alias{get_senseBox_Ids}
\title{List all senseBoxIds}
\usage{
get_senseBox_Ids(progress = FALSE)
}
\arguments{
\item{progress}{\link{logical}: Display the progress of the download?}
}
\value{
A \link{data.frame} with senseBoxIds and names of every senseBox
}
\description{
List all senseBoxIds
}
\section{Function version}{
 0.0.1
}

\examples{

\dontrun{
get_senseBox_Ids()
}

}
\author{
Johannes Friedrich
}
