% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senseBox_location.R
\name{get_senseBox_location}
\alias{get_senseBox_location}
\title{Get name, longitude and latitude of the submitted senseBoxId}
\usage{
get_senseBox_location(senseBoxId)
}
\arguments{
\item{senseBoxId}{\link{character} (\strong{required}): senseBoxId}
}
\value{
A \link{data.frame} with name, longtidue and latitude of the submitted senseBoxIds.
This \link{data.frame} can directly be used in \link[leaflet:leaflet]{leaflet::leaflet} to visualise the location of the senseBoxIds.
}
\description{
Get name, longitude and latitude of the submitted senseBoxId
}
\section{Function version}{
 0.0.1
}

\examples{

get_senseBox_location("593acaa66ccf3b00116deb0f")

}
\author{
Johannes Friedrich
}
