% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senseBox_sensor_Ids.R
\name{get_senseBox_sensor_Ids}
\alias{get_senseBox_sensor_Ids}
\title{Get sensor Ids from senseBox}
\usage{
get_senseBox_sensor_Ids(senseBoxId)
}
\arguments{
\item{senseBoxId}{\link{character} (\strong{required}): senseBoxId}
}
\value{
\link{list}
}
\description{
Get sensor Ids from senseBox
}
\section{Function version}{
 0.0.1
}

\examples{

##=====================================
## Example: get senosorIds from one sensBox
##=====================================

get_senseBox_sensor_Ids("593acaa66ccf3b00116deb0f")

}
\author{
Johannes Friedrich
}
