% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senseBox_sensor_info.R
\name{get_senseBox_sensor_info}
\alias{get_senseBox_sensor_info}
\title{Get info (Ids) from sensors of a senseBox}
\usage{
get_senseBox_sensor_info(senseBoxId)
}
\arguments{
\item{senseBoxId}{\link{character} (\strong{required}): senseBoxId}
}
\value{
\link{list}
}
\description{
Get info (Ids) from sensors of a senseBox
}
\section{Function version}{
 0.0.1
}

\examples{

##=========================================
## Example: get senosorIds from one sensBox
##=========================================

get_senseBox_sensor_info("593acaa66ccf3b00116deb0f")

}
\author{
Johannes Friedrich
}
