% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senseBox_stats.R
\name{get_senseBox_stats}
\alias{get_senseBox_stats}
\title{Get statistics about senseBox project}
\usage{
get_senseBox_stats()
}
\value{
A \link{data.frame} with "Number of senseBoxes",
"Number of Measurements" and "Number of measurements in last minute" from the senseBox API
}
\description{
Get statistics about senseBox project
}
\section{Function version}{
 0.0.1
}

\examples{

get_senseBox_stats()

}
\author{
Johannes Friedrich
}
