% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_senseBoxData.R
\name{parse_senseBoxData}
\alias{parse_senseBoxData}
\title{Construct a senseBox data.frame}
\usage{
parse_senseBoxData(boxdata)
}
\arguments{
\item{boxdata}{A named \code{list} containing the data from a senseBox}
}
\value{
A \code{data.frame}
}
\description{
Parses the fields of a \code{/boxes} response from the openSenseMap API
The code was originally developed by Norwin Roosen (https://github.com/noerw/opensensmapR)
}
\author{
Norwin Roosen, University of Muenster (Germany),
Johannes Friedrich, University of Bayreuth (Germany)
}
