% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senseBox-package.R
\docType{package}
\name{senseBox-package}
\alias{senseBox-package}
\title{senseBox}
\description{
A R API wrapper for the senseBox project. Download and analyse environmental data provided by https://sensebox.de/en/.
}
\details{
\tabular{ll}{Package: \tab senseBox\cr Type: \tab Package\cr Version:
\tab 0.0.1 \cr Date: \tab 2017-XX-XX \cr License: \tab GPL-3\cr }
}
\keyword{package}
