% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senseBox_Ids.R
\name{get_senseBox_Ids}
\alias{get_senseBox_Ids}
\title{List all senseBoxIds}
\usage{
get_senseBox_Ids(txtProgressBar = TRUE)
}
\arguments{
\item{txtProgressBar}{\link{logical}: Display a progress bar?}
}
\value{
\link{data.frame}
}
\description{
List all senseBoxIds
}
\section{Function version}{
 0.0.1
}

\examples{

##================================================
## Example: Get all available Ids from sensBox API
##================================================
\dontrun{
get_senseBox_Ids()
}
}
\author{
Johannes Friedrich
}
