% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senseBox_info.R
\name{get_senseBox_info}
\alias{get_senseBox_info}
\title{Get info about one senseBox.
You will receive information about a senseBox, e.g., the different sensors.}
\usage{
get_senseBox_info(senseBoxId, parallel = FALSE, tidy = FALSE)
}
\arguments{
\item{senseBoxId}{\link{character} or \link{vector} of \link{character} (\strong{required}): senseBoxId}

\item{parallel}{\link{logical} (\strong{optional}): Should the calculations be executed on multiple cores?
At least 4 cores are necessary to activate this feature.}

\item{tidy}{\link{logical} (\strong{optional}): Should the output be a tidy data frame?}
}
\value{
A \link{data.frame}
}
\description{
Get info about one senseBox.
You will receive information about a senseBox, e.g., the different sensors.
}
\section{Function version}{
 0.0.1
}

\examples{

##=====================================
## Example: get data from one senseBox
##=====================================

get_senseBox_info("593acaa66ccf3b00116deb0f")

}
\author{
Johannes Friedrich
}
