% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senseBox-package.R
\docType{package}
\name{senseBox-package}
\alias{senseBox-package}
\title{senseBox}
\description{
A collection of functions to illustrate the usage of smartphone sensors with the app HyperIMU
}
\details{
\tabular{ll}{Package: \tab HypeRIMU\cr Type: \tab Package\cr Version:
\tab 0.0.1 \cr Date: \tab 2017-XX-XX \cr License: \tab GPL-3\cr }
}
\keyword{package}
